fix: align alternative origin check with backend
